+++ NEWS.md
@@ -1,6 +1,7 @@
 # Version 0.1.3
 
 * Empty flag is invalid and throws an error (#3).
+* Allow verb arguments for sub-commands with a `command()` method to access these from within the scripts (#4).
 
 # Version 0.1.2